third_party/blink/web_tests/svg/dynamic-updates/SVGLinearGradientElement-dom-x2-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x2' attribute of the SVGLinearGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS linearGradientElement.getAttribute('x2') is "-100%"
PASS linearGradientElement.getAttribute('x2') is "100%"
PASS successfullyParsed is true

TEST COMPLETE
